cleanup scripts.js of code that des not need be there
