+++ styles/styles.css
@@ -939,7 +939,7 @@ main > div.browse-rail {
   display: none;
 }
 
-body[class^=browse-] main.browse-background-img {
+body[class^=browse-] main {
   background-image: url("/solutions/icons/BrowseDecorationMobile.svg");
   background-repeat: no-repeat;
   background-position: top right;
@@ -985,7 +985,7 @@ body[class^=browse-] .browse-filters-full-container .browse-filters-pagination, 
   .section.highlight.icon-block-container > div > h3 {
     text-align: left;
   }
-  body[class^=browse-] main.browse-background-img {
+  body[class^=browse-] main {
     background-image: url("/solutions/icons/BrowseDecoration.svg");
     background-size: contain;
   }
@@ -994,11 +994,6 @@ body[class^=browse-] .browse-filters-full-container .browse-filters-pagination, 
   }
 }
 @media (min-width: 900px) {
-  .docs main.rail-left-closed.rail-right-closed .section.content-section-first.content-section-last {
-    padding-right: 0;
-    margin-right: 0;
-    max-width: calc(100vw - 100px);
-  }
   .docs main.rail-left-closed.rail-right-closed .default-content-wrapper p {
     max-width: 95%;
   }
@@ -1042,11 +1037,6 @@ body[class^=browse-] .browse-filters-full-container .browse-filters-pagination, 
     width: 224px;
     height: 100%;
     padding: 0 0 0 32px;
-  }
-}
-@media (width >= 1440px) {
-  body[class^=browse-] main {
-    background-position: -36px 0;
   }
 }
 @keyframes placeholder-shimmer {
